regorganize improvements dialog
