+++ src/components/improvements/complexForm/improvementComplexFormList.test.tsx
@@ -1,6 +1,7 @@
 import { fireEvent, screen } from "@testing-library/react"
 import { describe, expect, it } from "vitest"
 
+import { SpendKarmaDialogProvider } from "#/components/improvements/spendKarmaDialogContext.tsx"
 import { RunnerDataStore } from "#/components/runner/sheet/runnerDataStore.ts"
 import { AttributeKey } from "#/system/attributeKey.ts"
 import { runnerDataFactory } from "#/system/runnerData.factory.ts"
@@ -8,7 +9,6 @@ import type { RunnerData } from "#/system/runnerData.ts"
 import { renderWithProviders } from "#testUtils/renderUtils.tsx"
 
 import { ImprovementComplexFormList } from "./improvementComplexFormList.tsx"
-import { SpendKarmaDialogProvider } from "./spendKarmaDialogContext.tsx"
 
 function renderList(updateRunnerData?: (sheet: RunnerData) => void) {
   return renderWithProviders(
