+++ src/components/improvements/skills/improvementKnowledgeSkillList.test.tsx
@@ -1,13 +1,13 @@
 import { fireEvent, screen } from "@testing-library/react"
 import { describe, expect, it } from "vitest"
 
+import { SpendKarmaDialogProvider } from "#/components/improvements/spendKarmaDialogContext.tsx"
 import { RunnerDataStore } from "#/components/runner/sheet/runnerDataStore.ts"
 import { runnerDataFactory } from "#/system/runnerData.factory.ts"
 import type { RunnerData } from "#/system/runnerData.ts"
 import { renderWithProviders } from "#testUtils/renderUtils.tsx"
 
 import { ImprovementKnowledgeSkillList } from "./improvementKnowledgeSkillList.tsx"
-import { SpendKarmaDialogProvider } from "./spendKarmaDialogContext.tsx"
 
 function renderList(updateRunnerData?: (sheet: RunnerData) => void) {
   return renderWithProviders(
