+++ src/components/improvements/spendKarmaSections.tsx
@@ -18,16 +18,16 @@ import { isTechnomancer } from "#/components/runner/technomancer/technomancerUti
 import type { AwakeningType } from "#/system/awakeningType.ts"
 import { isMagical } from "#/system/awakeningType.ts"
 
-import { ImprovementActiveSkillList } from "./improvementActiveSkillList.tsx"
-import { ImprovementAttributeList } from "./improvementAttributeList.tsx"
-import { ImprovementComplexFormList } from "./improvementComplexFormList.tsx"
-import { ImprovementInitiationList } from "./improvementInitiationList.tsx"
-import { ImprovementKnowledgeSkillList } from "./improvementKnowledgeSkillList.tsx"
-import { ImprovementLanguageSkillList } from "./improvementLanguageSkillList.tsx"
-import { ImprovementQualityList } from "./improvementQualityList.tsx"
-import { ImprovementSkillGroupList } from "./improvementSkillGroupList.tsx"
-import { ImprovementSpellList } from "./improvementSpellList.tsx"
-import { ImprovementSubmersionList } from "./improvementSubmersionList.tsx"
+import { ImprovementAttributeList } from "./attributes/improvementAttributeList.tsx"
+import { ImprovementComplexFormList } from "./complexForm/improvementComplexFormList.tsx"
+import { ImprovementInitiationList } from "./initiation/improvementInitiationList.tsx"
+import { ImprovementQualityList } from "./qualities/improvementQualityList.tsx"
+import { ImprovementActiveSkillList } from "./skills/improvementActiveSkillList.tsx"
+import { ImprovementKnowledgeSkillList } from "./skills/improvementKnowledgeSkillList.tsx"
+import { ImprovementLanguageSkillList } from "./skills/improvementLanguageSkillList.tsx"
+import { ImprovementSkillGroupList } from "./skills/improvementSkillGroupList.tsx"
+import { ImprovementSpellList } from "./spells/improvementSpellList.tsx"
+import { ImprovementSubmersionList } from "./submersion/improvementSubmersionList.tsx"
 
 export type SpendKarmaSectionKey =
   | "attribute"
